+++ src/streamtree/core/__init__.py
@@ -1,6 +1,6 @@
 """Core primitives: elements, components, render."""
 
-from streamtree.core.component import component, render
+from streamtree.core.component import component, render, render_app
 from streamtree.core.element import ComponentCall, Element, Fragment, fragment
 
 __all__ = [
@@ -10,4 +10,5 @@ __all__ = [
     "component",
     "fragment",
     "render",
+    "render_app",
 ]